Rename Project and Service to more accurate names
