--- frontend/src/Components/Project/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Project";
