+++ frontend/src/Components/ProjectCard/ProjectCard.scss
