--- frontend/src/Components/Service/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Service";
