+++ frontend/src/Components/index.js
@@ -2,9 +2,9 @@ export { default as CustomToolbar } from "./CustomToolbar";
 export { default as CustomFooter } from "./CustomFooter";
 export { default as PageHeader } from "./PageHeader";
 export { default as FooterList } from "./FooterList";
-export { default as Service } from "./Service";
 export { default as CustomServicesNav } from "./CustomServicesNav";
-export { default as Project } from "./Project";
+export { default as ProjectCard } from "./ProjectCard";
+export { default as ServiceCard } from "./ServiceCard";
 export { default as ContactForm } from "./ContactForm";
 export { default as CareersForm } from "./CareersForm";
 export { default as LanguageMenu } from "./LanguageMenu";
